# src/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ApexBridge REST API",
    "version": "1.0.0",
    "description": "High-yield crypto and investment portal REST API specification. Supports complete authentication, live wallet tracking, deposit/withdrawal lifecycles, and yield management.",
    "contact": {
      "name": "ApexBridge Engineering",
      "email": "support@apexbridge.io",
      "url": "https://api.apexbridge.io"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server (Local or Cloud)"
    },
    {
      "url": "https://api.apexbridge.io",
      "description": "ApexBridge Production Base URL"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token obtained from `/api/v1/auth/login` or `/api/v1/auth/signup`."
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "usr_8829104" },
          "name": { "type": "string", "example": "Alexander Vance" },
          "email": { "type": "string", "example": "alexander@apexbridge.com" },
          "avatar": { "type": "string", "example": "https://images.unsplash.com/photo-1534528741775-53994a69daeb" },
          "role": { "type": "string", "example": "investor" },
          "tier": { "type": "string", "example": "Tier 2 - Verified" },
          "balance": { "type": "number", "example": 48250.00 },
          "is2FAEnabled": { "type": "boolean", "example": true },
          "createdAt": { "type": "string", "example": "2026-08-30T10:00:00Z" }
        }
      }
    }
  },
  "paths": {
    "/api/v1/auth/signup": {
      "post": {
        "summary": "Register a new investor account",
        "tags": ["1. Authentication & Sessions"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["fullName", "email", "password"],
                "properties": {
                  "fullName": { "type": "string", "example": "Alexander Vance" },
                  "email": { "type": "string", "example": "alexander@apexbridge.com" },
                  "password": { "type": "string", "example": "SecurePassword123!" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "Authenticate existing credentials and start session",
        "tags": ["1. Authentication & Sessions"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "alexander@apexbridge.com" },
                  "password": { "type": "string", "example": "SecurePassword123!" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated session active",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "summary": "Retrieve current authenticated session and profile",
        "tags": ["1. Authentication & Sessions"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/deposits/{id}/status": {
      "patch": {
        "summary": "Admin approval or rejection of user deposits",
        "tags": ["9. Administrative & Back-Office Operations"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "tx_dep_99401" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { "type": "string", "enum": ["approved", "rejected", "pending"], "example": "approved" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deposit status updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "transactionId": { "type": "string", "example": "tx_dep_99401" },
                    "status": { "type": "string", "example": "approved" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/withdrawals/{id}/status": {
      "patch": {
        "summary": "Admin processing or rejection of withdrawal requests",
        "tags": ["9. Administrative & Back-Office Operations"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "tx_wdr_10294" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { "type": "string", "enum": ["processed", "approved", "rejected", "pending"], "example": "processed" },
                  "txHash": { "type": "string", "example": "0x78a1bc92048592019485091823" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Withdrawal status updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "transactionId": { "type": "string", "example": "tx_wdr_10294" },
                    "status": { "type": "string", "example": "processed" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/plans/{id}": {
      "put": {
        "summary": "Update investment plan parameters, returns, and minimums",
        "tags": ["9. Administrative & Back-Office Operations"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "starter" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roi": { "type": "string", "example": "18%" },
                  "minAmount": { "type": "number", "example": 1000 },
                  "maxAmount": { "type": "number", "example": 10000 },
                  "feeRate": { "type": "number", "example": 0.08 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Plan configuration updated" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}